backend/upload-property-photo/requirements.txt
boto3>=1.26.0
Pillow>=10.0.0
requests>=2.31.0